Make it work with Lua 5.4+
