+++ src/websocket/bit.lua
@@ -1,31 +1,18 @@
-local bit = {}
+local has_bit33, bit = pcall(require,'bit32')
+local support_bitwise = loadstring("__xxx = 1 << 1")
 
-function bit.lshift(a, b)
-  return a<<b
+-- For Lua 5.3.5+, 5.4+
+if support_bitwise then
+  bit = require"./bitpoly"
+  return bit
 end
 
-function bit.rshift(a, b)
-  return a>>b
+if has_bit32 then
+  -- lua 5.2 / bit32 library
+  bit.rol = bit.lrotate
+  bit.ror = bit.rrotate
+  return bit
+else
+  -- luajit / lua 5.1 + luabitop
+  return require'bit'
 end
-
-function bit.band(a, b)
-  return a&b
-end
-
-function bit.bor(a, b)
-  return a|b
-end
-
-function bit.bxor(a, b)
-  return a~b
-end
-
-function bit.rol(a, b)
-  return (a << b) | (a >> (32 - b)) & ~(-1 << b);
-end
-
-function bit.bnot(a)
-  return ~a
-end
-
-return bi
